==> 1 - Create Edges File.ipynb <==
import pandas as pd
import numpy as np

# Read in and show the the med_full_final_melted.csv file
med_claims = pd.read_csv('Data/med_full_final_melted.csv')
med_claims = med_claims.drop('Unnamed: 0', axis = 1)
med_claims

# Create a list of the unique Member Life IDs
unique_ids = np.unique(med_claims['Member Life ID'])

edges_dict = {}

for current_id in unique_ids:
    # Slice the medical claims dataframe to get only the rows where Member Life ID equals the current_id
    current_slice = med_claims.loc[med_claims['Member Life ID'] == current_id]

    current_base_codes = np.sort(np.unique(current_slice['Base Code']))

    # Create a list of edges to include in the edges_dict dictionary.
    current_edges = []
    for node_1 in current_base_codes:
        for node_2 in current_base_codes:
            if node_1 != node_2 and (node_2, node_1) not in current_edges:
                current_edges.append((node_1, node_2))

    # Loop through the current_edges list and update the edges_dict dictionary appropriately.
    for edge in current_edges:
        node_1 = edge[0]
        node_2 = edge[1]

        if edge in edges_dict.keys():
            edges_dict[edge] = edges_dict[edge] + 1
        else:
            edges_dict[edge] = 1


edges_dict

# Create a dataframe for the edges and weights and then write a csv
final_edges_df = pd.DataFrame({'Edge' : edges_dict.keys(),'Weight':edges_dict.values()})
final_edges_df['Source'] = final_edges_df['Edge'].str[0]
final_edges_df['Target'] = final_edges_df['Edge'].str[1]

# Load in the code-description pairs
code_description_pairs = pd.read_table('Data/code_description_pairs.txt', header=None)
code_description_pairs = code_description_pairs.rename({0:'Code', 1:'Description'}, axis=1)

# Join in the Source Descriptions
final_edges_df = pd.merge(final_edges_df, code_description_pairs, how = 'left', left_on = 'Source', right_on = 'Code')
final_edges_df = final_edges_df.rename({'Description': 'Source Description'}, axis=1)
final_edges_df = final_edges_df.drop('Code', axis = 1)

# Join in the Target Descriptions
final_edges_df = pd.merge(final_edges_df, code_description_pairs, how = 'left', left_on = 'Target', right_on = 'Code')
final_edges_df = final_edges_df.rename({'Description': 'Target Description'}, axis=1)
final_edges_df = final_edges_df.drop('Code', axis = 1)

# Sort the final_edges_df dataframe
final_edges_df = final_edges_df.sort_values(by = ['Weight', 'Source', 'Target'], ascending = [False, True, True]).reset_index(drop=True)
final_edges_df

# Write the final_edges_df to csv
final_edges_df.to_csv('edges.csv')
