==> 3 - Final Dataset Changes.ipynb <==
# Import required libraries
import pandas as pd
import numpy as np

#Creating a function to automatically identify possible predictor ICD10 codes. Use this function in other notebooks.
def get_predictor_codes(edges, exclusions, input_code, limit = 20):
    filtered_edges = edges.loc[edges['Edge'].str.contains(input_code)].sort_values(by = 'Weight', ascending = False).reset_index(drop=True)

    # If there are exclusions remove edges where the other base code is in the exclusion list 
    if len(exclusions) != 0:
        filtered_edges = filtered_edges.loc[~filtered_edges['Edge'].str.contains('|'.join(exclusions))].reset_index(drop=True)



    # Return the other base code of the pair
    other_codes = [''] * limit
    other_desc = [''] * limit
    weights = [''] * limit

    for i in range(limit):
        current_source = filtered_edges['Source'].iloc[i]
        current_source_desc = filtered_edges['Source Description'].iloc[i]
        current_target = filtered_edges['Target'].iloc[i]
        current_target_desc = filtered_edges['Target Description'].iloc[i]

        if current_source == input_code:
            other_codes[i] = current_target
            other_desc[i] = current_target_desc
        else:
            other_codes[i] = current_source
            other_desc[i] = current_source_desc
        weights[i] = filtered_edges['Weight'].iloc[i]
    return (other_codes, other_desc, weights)


# Importing the formatted dataset and the exclusions dataset
all_data = pd.read_csv('Data/formatted_data_all_codes.csv')
all_data = all_data.drop('Unnamed: 0', axis=1)

# Remove member life IDs that have conflicting Biological Genders recorded.
# Get the Member Life IDs with two or more rows
gender_df = all_data[['Member Life ID', 'Biological Gender']]

gender_df = gender_df.loc[gender_df['Member Life ID'].duplicated(keep=False)].drop_duplicates()

ids_with_two_genders = gender_df[gender_df['Member Life ID'].duplicated(keep=False)]['Member Life ID'].drop_duplicates().to_list()

all_data = all_data[~all_data['Member Life ID'].isin(ids_with_two_genders)]
all_data = all_data.reset_index(drop=True)

all_data

# Create an all_data_with_ages dataframe
# Rows with Ages Only (Assuming that the Header Service From Date is the one to use)
all_data_with_ages = all_data.loc[pd.notna(all_data['Birth Date'])]
all_data_with_ages = all_data_with_ages.drop('Line Service From Date', axis = 1)
all_data_with_ages[['Header Service From Date', 'Birth Date']] = all_data_with_ages[['Header Service From Date', 'Birth Date']].apply(pd.to_datetime)
all_data_with_ages['Age'] = np.floor((all_data_with_ages['Header Service From Date'] - all_data_with_ages['Birth Date']).dt.days/365.25)
all_data_with_ages = all_data_with_ages.drop(['Header Service From Date', 'Birth Date'], axis=1)
all_data_with_ages = all_data_with_ages.drop_duplicates()
all_data_with_ages = all_data_with_ages.reset_index(drop=True)

# Removing Member Life IDs that have more than one age
# Get the Member Life IDs with two or more rows
age_df = all_data_with_ages[['Member Life ID', 'Age']]

age_df = age_df.loc[age_df['Member Life ID'].duplicated(keep=False)].drop_duplicates()

ids_with_two_ages = age_df[age_df['Member Life ID'].duplicated(keep=False)]['Member Life ID'].drop_duplicates().to_list()

all_data_with_ages = all_data_with_ages[~all_data_with_ages['Member Life ID'].isin(ids_with_two_ages)]
all_data_with_ages = all_data_with_ages.reset_index(drop=True)

all_data_with_ages

# Write the all_data_with_ages dataframe to a csv
all_data_with_ages.to_csv('all_data_with_ages.csv')
