==> 7 - Evaluation Tree Model With HyperParam With Difference threshold.ipynb <==
# Import required libraries
import pandas as pd
from sklearn.metrics import roc_curve, auc, plot_roc_curve, roc_auc_score, accuracy_score, precision_recall_curve
from sklearn.model_selection import cross_validate, train_test_split, GridSearchCV
from sklearn.linear_model import LogisticRegressionCV
from sklearn.ensemble import RandomForestClassifier
import numpy as np
import matplotlib.pyplot as plt
import seaborn
from sklearn.model_selection import cross_val_score
from sklearn.metrics import classification_report, confusion_matrix
from random_forest_functions import *
import sys

random_seed = 51

medical_claims = pd.read_csv('Data/med_full_final_melted.csv')
medical_claims

# Load in the pharmacy claims data
rx_claims = pd.read_csv('Data/rx_full_final.csv')

# Load in the edges file and exclusions file
edges = pd.read_csv('Data/edges.csv')
exclusions = pd.read_csv('Data/exclusions.csv')
exclusion_list = exclusions['Base Code'].to_list()

all_data_with_ages_without_timebound = pd.read_csv('./Data/all_data_with_ages.csv')

E11_random_forest_hyper_params_with_ages_without_timebound={'max_depth': 8, 'n_estimators': 300}
C18_random_forest_hyper_params_with_ages_without_timebound={'max_depth': 2, 'n_estimators': 100}
C50_random_forest_hyper_params_with_ages_without_timebound={'max_depth': 8, 'n_estimators': 100}
I10_random_forest_hyper_params_with_ages_without_timebound={'max_depth': 8, 'n_estimators': 400}
I25_random_forest_hyper_params_with_ages_without_timebound={'max_depth': 8, 'n_estimators': 300}
N18_random_forest_hyper_params_with_ages_without_timebound={'max_depth': 6, 'n_estimators': 400}

E11_random_forest_hyper_params_with_ages_with_timebound={'max_depth': 2, 'n_estimators': 100}
C18_random_forest_hyper_params_with_ages_with_timebound={'max_depth': 2, 'n_estimators': 100}
C50_random_forest_hyper_params_with_ages_with_timebound={'max_depth': 8, 'n_estimators': 300}
I10_random_forest_hyper_params_with_ages_with_timebound={'max_depth': 8, 'n_estimators': 300}
I25_random_forest_hyper_params_with_ages_with_timebound={'max_depth': 8, 'n_estimators': 100}
N18_random_forest_hyper_params_with_ages_with_timebound={'max_depth': 6, 'n_estimators': 300}

# Model for E11 with Age and Without the Timebound
E11_with_age_without_Timebound_model, E11_with_age_without_Timebound_model_Pred_prob, E11_with_age_without_Timebound_model_pred_codes, E11_with_age_without_Timebound_model_train_test =gen_rf_model(all_data_with_ages_without_timebound, 'E11', edges, exclusion_list, E11_random_forest_hyper_params_with_ages_without_timebound)
get_accuracy_sensitivity_for_all_threshold(E11_with_age_without_Timebound_model, E11_with_age_without_Timebound_model_train_test, E11_with_age_without_Timebound_model_Pred_prob, 'E11', input_type='without' )

print("\n \n \n")

# Model for E11 with Age and With the Timebound
E11_with_age_with_Timebound_model, E11_with_age_with_Timebound_model_Pred_prob, E11_with_age_with_Timebound_model_pred_codes, E11_with_age_with_Timebound_model_train_test =gen_rf_model_timebound(medical_claims, rx_claims, 'E11', edges, exclusion_list, E11_random_forest_hyper_params_with_ages_with_timebound)
get_accuracy_sensitivity_for_all_threshold(E11_with_age_with_Timebound_model, E11_with_age_with_Timebound_model_train_test, E11_with_age_with_Timebound_model_Pred_prob, 'E11', input_type='with' )

# Model for C18 with Age and Without the Timebound
C18_with_age_without_Timebound_model, C18_with_age_without_Timebound_model_Pred_prob, C18_with_age_without_Timebound_model_pred_codes, C18_with_age_without_Timebound_model_train_test =gen_rf_model(all_data_with_ages_without_timebound, 'C18', edges, exclusion_list, C18_random_forest_hyper_params_with_ages_without_timebound)
get_accuracy_sensitivity_for_all_threshold(C18_with_age_without_Timebound_model, C18_with_age_without_Timebound_model_train_test, C18_with_age_without_Timebound_model_Pred_prob, 'C18', input_type='without' )

print("\n \n \n")

# Model for C18 with Age and With the Timebound
C18_with_age_with_Timebound_model, C18_with_age_with_Timebound_model_Pred_prob, C18_with_age_with_Timebound_model_pred_codes, C18_with_age_with_Timebound_model_train_test =gen_rf_model_timebound(medical_claims, rx_claims, 'C18', edges, exclusion_list, C18_random_forest_hyper_params_with_ages_with_timebound)
get_accuracy_sensitivity_for_all_threshold(C18_with_age_with_Timebound_model, C18_with_age_with_Timebound_model_train_test, C18_with_age_with_Timebound_model_Pred_prob, 'C18', input_type='with' )

# Model for C50 with Age and Without the Timebound
C50_with_age_without_Timebound_model, C50_with_age_without_Timebound_model_Pred_prob, C50_with_age_without_Timebound_model_pred_codes, C50_with_age_without_Timebound_model_train_test =gen_rf_model(all_data_with_ages_without_timebound, 'C50', edges, exclusion_list, C50_random_forest_hyper_params_with_ages_without_timebound)
get_accuracy_sensitivity_for_all_threshold(C50_with_age_without_Timebound_model, C50_with_age_without_Timebound_model_train_test, C50_with_age_without_Timebound_model_Pred_prob, 'C50', input_type='without' )

print("\n \n \n")

# Model for C50 with Age and With the Timebound
C50_with_age_with_Timebound_model, C50_with_age_with_Timebound_model_Pred_prob, C50_with_age_with_Timebound_model_pred_codes, C50_with_age_with_Timebound_model_train_test =gen_rf_model_timebound(medical_claims, rx_claims, 'C50', edges, exclusion_list, C50_random_forest_hyper_params_with_ages_with_timebound)
get_accuracy_sensitivity_for_all_threshold(C50_with_age_with_Timebound_model, C50_with_age_with_Timebound_model_train_test, C50_with_age_with_Timebound_model_Pred_prob, 'C50', input_type='with' )

# Model for I10 with Age and Without the Timebound
I10_with_age_without_Timebound_model, I10_with_age_without_Timebound_model_Pred_prob, I10_with_age_without_Timebound_model_pred_codes, I10_with_age_without_Timebound_model_train_test =gen_rf_model(all_data_with_ages_without_timebound, 'I10', edges, exclusion_list, I10_random_forest_hyper_params_with_ages_without_timebound)
get_accuracy_sensitivity_for_all_threshold(I10_with_age_without_Timebound_model, I10_with_age_without_Timebound_model_train_test, I10_with_age_without_Timebound_model_Pred_prob, 'I10', input_type='without' )

print("\n \n \n")

# Model for I10 with Age and With the Timebound
I10_with_age_with_Timebound_model, I10_with_age_with_Timebound_model_Pred_prob, I10_with_age_with_Timebound_model_pred_codes, I10_with_age_with_Timebound_model_train_test =gen_rf_model_timebound(medical_claims, rx_claims, 'I10', edges, exclusion_list, I10_random_forest_hyper_params_with_ages_with_timebound)
get_accuracy_sensitivity_for_all_threshold(I10_with_age_with_Timebound_model, I10_with_age_with_Timebound_model_train_test, I10_with_age_with_Timebound_model_Pred_prob, 'I10', input_type='with' )

# Model for I25 with Age and Without the Timebound
I25_with_age_without_Timebound_model, I25_with_age_without_Timebound_model_Pred_prob, I25_with_age_without_Timebound_model_pred_codes, I25_with_age_without_Timebound_model_train_test =gen_rf_model(all_data_with_ages_without_timebound, 'I25', edges, exclusion_list, I25_random_forest_hyper_params_with_ages_without_timebound)
get_accuracy_sensitivity_for_all_threshold(I25_with_age_without_Timebound_model, I25_with_age_without_Timebound_model_train_test, I25_with_age_without_Timebound_model_Pred_prob, 'I25', input_type='without' )

print("\n \n \n")

# Model for I25 with Age and With the Timebound
I25_with_age_with_Timebound_model, I25_with_age_with_Timebound_model_Pred_prob, I25_with_age_with_Timebound_model_pred_codes, I25_with_age_with_Timebound_model_train_test =gen_rf_model_timebound(medical_claims, rx_claims, 'I25', edges, exclusion_list, I25_random_forest_hyper_params_with_ages_with_timebound)
get_accuracy_sensitivity_for_all_threshold(I25_with_age_with_Timebound_model, I25_with_age_with_Timebound_model_train_test, I25_with_age_with_Timebound_model_Pred_prob, 'I25', input_type='with' )

# Model for N18 with Age and Without the Timebound
N18_with_age_without_Timebound_model, N18_with_age_without_Timebound_model_Pred_prob, N18_with_age_without_Timebound_model_pred_codes, N18_with_age_without_Timebound_model_train_test =gen_rf_model(all_data_with_ages_without_timebound, 'N18', edges, exclusion_list, N18_random_forest_hyper_params_with_ages_without_timebound)
get_accuracy_sensitivity_for_all_threshold(N18_with_age_without_Timebound_model, N18_with_age_without_Timebound_model_train_test, N18_with_age_without_Timebound_model_Pred_prob, 'N18', input_type='without' )

print("\n \n \n")

# Model for N18 with Age and With the Timebound
N18_with_age_with_Timebound_model, N18_with_age_with_Timebound_model_Pred_prob, N18_with_age_with_Timebound_model_pred_codes, N18_with_age_with_Timebound_model_train_test =gen_rf_model_timebound(medical_claims, rx_claims, 'N18', edges, exclusion_list, N18_random_forest_hyper_params_with_ages_with_timebound)
get_accuracy_sensitivity_for_all_threshold(N18_with_age_with_Timebound_model, N18_with_age_with_Timebound_model_train_test, N18_with_age_with_Timebound_model_Pred_prob, 'N18', input_type='with' )
